centralize config (env var) access
